OptiXRenderer: Disable gamma correcting input data to the AI denoiser by default
Gamma correcting the color values means that the colors are no longer in a linear space and interpolating these non-linear colors in dark, noisy regions leads to overdarkening. Disabling gamma correction enables the AI denoiser (that is probably essentially linear) to correctly linearly interpolate the colors. Since OptiX 6.5 this doesn't seem to create artifacts in non-dark and noisy regions.