OptiXRenderer: Remove restriction on low value PDF and wi cos_theta. Instead test if directions are grazing angle and evaluate contribution to zero
The PDF computation has gotten more robust since the limitation was introduced and it was also non-symmetric, as the restriction was only when sampling, not when evaluation samples.
The same applies to the restriction to wi cos_theta, but additionally it was non-symmetric with respect to wi/wo, as the restriction only applied to the sampled direction and not the input direction.
Introduce grazing angle tests that ensures that grazing angle samples evaluate to black.