OptiXRenderer: Move renderer ID management to OptiXRenderer
In the theoretical case where we had multiple OptiX adaptors, they can't all set their own renderer ID, so it must be managed by the optix renderer.
The optix renderer is also the one actually using the ID, where as the Adaptor is just forwarding the ID to the renderer implementation, which again suggests that Adaptor is the wrong place to store it.