OptiXRenderer: Remove evaluate and PDF methods from shading models
They were only effectively used when validating that they worked, but that really boiled down to a lot of duplicated code and additional complexity.
The methods still exist on the individual BSDFs to have single-responsibility evaluate and PDF implementations that can validate the combined evaluate_with_PDF, which can contain shared computations and be more complex.