DX11Renderer: Correct environment map view direction computation
Strictly speaking this commit changes nothing in the environment map shader, as the view_to_world_matrix currently is the transpose of the world_to_view_matrix. But this change means that the vector/matrix multiplications are consistent with the multiplications in the other shaders, and if shearing is ever supported than it will just work for environments.
As the camera world position equals the translation in the view_to_world_matrix it has been removed from the scene constants and instead fetched from the view_to_world_matrix