OptiXRenderer: Stabilize primary ray direction
The previous ray direction was computed by subtracting two world positions. However, this became numerically unstable if camera was positioned such that the camera position was much further from the origin than the far plane. With the new compuation the direction only depends on the inverse projection matrix and the rotation of the view to world, making it independent of any positioning.