Bifrost: Fix assertion thrown in create_and_destroy_notifications LightSourceTest
THe issue was that when comparing a specific light type to to a light source ID, the light source ID would get cast to the light type, which could trigger an assert if the light source was of a different type. The fix was to use the more general light source equality comparison, as all they do is check the ID so the specific type wasn't needed.