Bifrost: Delay asset deletion until changes are reset
This avoids the issue where an asset with an old incarnation count is returned from the change set, and allows the components using the asset to handle it's dependencies. For example it might be useful to know which image a texture referenced, which isn't available if the property has been overwritten.