MaterialPrecomputations: Fix issue where dielectric hemispheres could be severely undersampled
The fix is to draw more samples until the minimum sample requirement is met.
This also makes the sample count a power of two, which is optimal for QMC random number generators.

Sampling useing the PracticalScrambledSobol RNG seems to converge to something else than Sample02, which is why the test thresholds get worse. PracticalScrambledSobol is newer though and used in PBRT, so we'll stick with that and potentially change the tests later.